+++ src/constants.ts
@@ -22,6 +22,7 @@ export const RENAME_CONNECTOR = `${COMMAND_PREFIX}.connector.rename`;
 export const DELETE_CONNECTOR = `${COMMAND_PREFIX}.connector.delete`;
 export const UPLOAD_CONNECTOR = `${COMMAND_PREFIX}.connector.upload`;
 export const DEPLOY_CONNECTOR = `${COMMAND_PREFIX}.connector.deploy`;
+export const INIT_CONNECTOR = `${COMMAND_PREFIX}.connector.init`;
 
 export const COPY_ID_CONNECTOR = `${COMMAND_PREFIX}.copy-id`;
 
@@ -30,6 +31,7 @@ export const RENAME_CUSTOMIZER = `${COMMAND_PREFIX}.customizer.rename`;
 export const DELETE_CUSTOMIZER = `${COMMAND_PREFIX}.customizer.delete`;
 export const UPLOAD_CUSTOMIZER = `${COMMAND_PREFIX}.customizer.upload`;
 export const DEPLOY_CUSTOMIZER = `${COMMAND_PREFIX}.customizer.deploy`;
+export const INIT_CUSTOMIZER = `${COMMAND_PREFIX}.customizer.init`;
 
 export const LINK_CUSTOMIZER_INSTANCE = `${COMMAND_PREFIX}.link`;
 export const UNLINK_CUSTOMIZER_INSTANCE = `${COMMAND_PREFIX}.unlink`;